+++ rust/net/chat/src/api/keytrans.rs
@@ -4,19 +4,20 @@
 //
 
 use std::borrow::Cow;
-use std::cmp::Ordering;
 use std::collections::{BTreeSet, HashMap};
 use std::fmt::{Debug, Formatter};
 use std::future::Future;
 use std::time::SystemTime;
 
 use async_trait::async_trait;
+use futures_util::future::try_join_all;
 use libsignal_core::{Aci, E164};
 use libsignal_keytrans::{
     AccountData, ChatDistinguishedResponse, ChatMonitorResponse, ChatSearchResponse,
     CondensedTreeSearchResponse, FullSearchResponse, FullTreeHead, KeyTransparency, LastTreeHead,
     LocalStateUpdate, MonitorContext, MonitorKey, MonitorProof, MonitorRequest, MonitorResponse,
     MonitoringData, SearchContext, SearchStateUpdate, SlimSearchRequest, VerifiedSearchResult,
+    Versioned,
 };
 use libsignal_net::env::KeyTransConfig;
 use libsignal_protocol::PublicKey;
@@ -27,7 +28,10 @@ const SEARCH_KEY_PREFIX_ACI: &[u8] = b"a";
 const SEARCH_KEY_PREFIX_E164: &[u8] = b"n";
 const SEARCH_KEY_PREFIX_USERNAME_HASH: &[u8] = b"u";
 
+const MAXIMUM_ALLOWED_VERSION_DELTA: u32 = 10;
+
 #[derive(Debug, thiserror::Error, displaydoc::Display)]
+#[cfg_attr(test, derive(Clone))]
 #[ignore_extra_doc_attributes]
 pub enum Error {
     /// Verification failed: {0}
@@ -42,10 +46,10 @@ pub enum Error {
 pub trait LowLevelChatApi {
     async fn search(
         &self,
-        aci: &Aci,
+        aci: Versioned<&Aci>,
         aci_identity_key: &PublicKey,
-        e164: Option<&(E164, Vec<u8>)>,
-        username_hash: Option<&UsernameHash<'_>>,
+        e164: Option<Versioned<&(E164, Vec<u8>)>>,
+        username_hash: Option<Versioned<&UsernameHash<'_>>>,
         stored_account_data: Option<&AccountData>,
         distinguished_tree_head: &LastTreeHead,
     ) -> Result<Vec<u8>, RequestError<Error>>; // Expected to be a SearchResponse proto
@@ -326,10 +330,10 @@ pub struct KeyTransparencyClient<'a> {
 pub trait UnauthenticatedChatApi {
     fn search(
         &self,
-        aci: &Aci,
+        aci: Versioned<&Aci>,
         aci_identity_key: &PublicKey,
-        e164: Option<(E164, Vec<u8>)>,
-        username_hash: Option<UsernameHash<'_>>,
+        e164: Option<Versioned<(E164, Vec<u8>)>>,
+        username_hash: Option<Versioned<UsernameHash<'_>>>,
         stored_account_data: Option<AccountData>,
         distinguished_tree_head: &LastTreeHead,
     ) -> impl Future<Output = Result<MaybePartial<AccountData>, RequestError<Error>>> + Send;
@@ -355,6 +359,47 @@ pub enum MonitorMode {
     MonitorOther,
 }
 
+#[derive(Debug, Clone, PartialEq)]
+struct SearchVersions {
+    aci: Option<u32>,
+    e164: Option<u32>,
+    username_hash: Option<u32>,
+}
+
+impl SearchVersions {
+    fn from_account_data(account_data: &AccountData) -> Self {
+        let AccountData {
+            aci,
+            e164,
+            username_hash,
+            ..
+        } = account_data;
+        Self {
+            aci: Some(aci.greatest_version()),
+            e164: e164.as_ref().map(|x| x.greatest_version()),
+            username_hash: username_hash.as_ref().map(|x| x.greatest_version()),
+        }
+    }
+
+    fn subtract(&self, other: &Self) -> Self {
+        fn opt_sub<T>(lhs: &T, rhs: &T, f: impl Fn(&T) -> Option<u32>) -> Option<u32> {
+            Some(f(lhs)? - f(rhs)?)
+        }
+        Self {
+            aci: opt_sub(self, other, |x| x.aci),
+            e164: opt_sub(self, other, |x| x.e164),
+            username_hash: opt_sub(self, other, |x| x.username_hash),
+        }
+    }
+
+    fn maximum_version(&self) -> Option<u32> {
+        [self.aci, self.e164, self.username_hash]
+            .into_iter()
+            .flatten()
+            .max()
+    }
+}
+
 pub async fn monitor_and_search(
     kt: &impl UnauthenticatedChatApi,
     aci: &Aci,
@@ -375,58 +420,160 @@ pub async fn monitor_and_search(
         )
         .await?;
 
+    let stored_versions = SearchVersions::from_account_data(&stored_account_data);
+    let updated_versions = SearchVersions::from_account_data(&updated_account_data);
+
     // Call to `monitor` guarantees that the optionality of E.164 and username hash data
     // will match between `stored_account_data` and `updated_account_data`. Meaning, they will
     // either both be Some() or both None.
-    let version_changed = has_version_changed_between(&stored_account_data, &updated_account_data);
 
-    // In case of self-monitoring, it is an error to detect a version change.
-    if version_changed && matches!(mode, MonitorMode::MonitorSelf) {
-        return Err(RequestError::Other(
-            libsignal_keytrans::Error::VerificationFailed(
-                "version change detected while self-monitoring".to_string(),
+    let final_account_data = match mode {
+        MonitorMode::MonitorSelf => {
+            // In case of self-monitoring, detecting a version change requires a fall
+            // back to search-with-version for all intermediate versions of all keys.
+            // It is important to only bump version of one search key at a time (aci,
+            // e164, or username_hash), though, and leave others at "baseline" level.
+            // The baseline is provided by search key's greatest known version from
+            // stored_account_data.
+
+            // SearchParameters that correspond to the stored_account_data, which is
+            // considered to be valid.
+            // Unwrapping stored versions is safe because monitor call above checks
+            // the "nullability mismatch" already.
+            let baseline_search_params = SearchParameters {
+                aci: Versioned::new(
+                    aci,
+                    stored_versions.aci.expect("ACI version must be present"),
+                ),
+                aci_identity_key,
+                e164: e164.as_ref().map(|e164| {
+                    Versioned::new(e164, stored_versions.e164.expect("checked earlier"))
+                }),
+                username_hash: username_hash.as_ref().map(|hash| {
+                    Versioned::new(
+                        hash,
+                        stored_versions.username_hash.expect("checked earlier"),
+                    )
+                }),
+                stored_account_data: stored_account_data.clone(),
+                distinguished_tree_head,
+            };
+
+            let version_delta = updated_versions.subtract(&stored_versions);
+
+            if version_delta
+                .maximum_version()
+                .is_some_and(|max_version| max_version > MAXIMUM_ALLOWED_VERSION_DELTA)
+            {
+                return Err(RequestError::Other(Error::VerificationFailed(
+                    libsignal_keytrans::Error::BadData("version changed by too much".to_string()),
+                )));
+            }
+
+            let all_search_params = {
+                let mut result = Vec::new();
+
+                if let Some(delta) = version_delta.aci {
+                    let baseline = stored_versions.aci.expect("must be present");
+                    for increment in 1..=delta {
+                        let mut params = baseline_search_params.clone();
+                        params.aci.version = Some(baseline + increment);
+                        // Don't search for other keys as an optimization
+                        params.e164 = None;
+                        params.username_hash = None;
+                        result.push(params);
+                    }
+                }
+
+                if let Some(delta) = version_delta.e164 {
+                    let baseline = stored_versions.e164.expect("checked by monitor call");
+                    for increment in 1..=delta {
+                        let mut params = baseline_search_params.clone();
+                        params
+                            .e164
+                            .as_mut()
+                            .expect("checked by monitor call")
+                            .version = Some(baseline + increment);
+                        // Don't search for other keys as an optimization
+                        params.username_hash = None;
+                        result.push(params);
+                    }
+                }
+
+                if let Some(delta) = version_delta.username_hash {
+                    let baseline = stored_versions
+                        .username_hash
+                        .expect("checked by monitor call");
+                    for increment in 1..=delta {
+                        let mut params = baseline_search_params.clone();
+                        params
+                            .username_hash
+                            .as_mut()
+                            .expect("checked by monitor call")
+                            .version = Some(baseline + increment);
+                        // Don't search for other keys as an optimization
+                        params.e164 = None;
+                        result.push(params);
+                    }
+                }
+                result
+            };
+
+            let futures = all_search_params
+                .into_iter()
+                .map(|params| search_with_params(kt, params));
+            try_join_all(futures).await?;
+            // If all the version specific searches succeed it is OK to accept the
+            // account data monitor returned.
+            updated_account_data.into()
+        }
+        MonitorMode::MonitorOther if updated_versions != stored_versions => {
+            kt.search(
+                Versioned::from(aci),
+                aci_identity_key,
+                e164.map(Versioned::from),
+                username_hash.map(Versioned::from),
+                Some(stored_account_data),
+                distinguished_tree_head,
             )
-            .into(),
-        ));
-    }
-
-    let final_account_data = if version_changed {
-        kt.search(
-            aci,
-            aci_identity_key,
-            e164,
-            username_hash,
-            Some(stored_account_data),
-            distinguished_tree_head,
-        )
-        .await?
-    } else {
-        updated_account_data.into()
+            .await?
+        }
+        MonitorMode::MonitorOther => updated_account_data.into(),
     };
     Ok(final_account_data)
 }
 
-fn cmp_by_key<T, K: Ord>(lhs: &T, rhs: &T, get_key: impl Fn(&T) -> K) -> Ordering {
-    get_key(lhs).cmp(&get_key(rhs))
+#[derive(Clone)]
+struct SearchParameters<'a> {
+    aci: Versioned<&'a Aci>,
+    aci_identity_key: &'a PublicKey,
+    e164: Option<Versioned<&'a (E164, Vec<u8>)>>,
+    username_hash: Option<Versioned<&'a UsernameHash<'a>>>,
+    stored_account_data: AccountData,
+    distinguished_tree_head: &'a LastTreeHead,
 }
 
-/// Compare the account data fields between stored and received from monitor call,
-/// and decide whether there is a newer version of search key in the tree, in which
-/// case search request needs to be performed.
-fn has_version_changed_between(stored: &AccountData, updated: &AccountData) -> bool {
-    let aci_version = |acc_data: &AccountData| Some(acc_data.aci.greatest_version());
-    let e164_version =
-        |acc_data: &AccountData| acc_data.e164.as_ref().map(|md| md.greatest_version());
-    let username_hash_version = |acc_data: &AccountData| {
-        acc_data
-            .username_hash
-            .as_ref()
-            .map(|md| md.greatest_version())
-    };
-
-    cmp_by_key(stored, updated, aci_version) == Ordering::Less
-        || cmp_by_key(stored, updated, e164_version) == Ordering::Less
-        || cmp_by_key(stored, updated, username_hash_version) == Ordering::Less
+async fn search_with_params(
+    kt: &impl UnauthenticatedChatApi,
+    params: SearchParameters<'_>,
+) -> Result<MaybePartial<AccountData>, RequestError<Error>> {
+    let SearchParameters {
+        aci,
+        aci_identity_key,
+        e164,
+        username_hash,
+        stored_account_data,
+        distinguished_tree_head,
+    } = params;
+    kt.search(
+        aci,
+        aci_identity_key,
+        e164.map(|x| x.cloned()),
+        username_hash.map(|x| x.cloned()),
+        Some(stored_account_data),
+        distinguished_tree_head,
+    )
+    .await
 }
 
 impl<'a> KeyTransparencyClient<'a> {
@@ -443,20 +590,20 @@ impl<'a> KeyTransparencyClient<'a> {
 impl UnauthenticatedChatApi for KeyTransparencyClient<'_> {
     async fn search(
         &self,
-        aci: &Aci,
+        aci: Versioned<&Aci>,
         aci_identity_key: &PublicKey,
-        e164: Option<(E164, Vec<u8>)>,
-        username_hash: Option<UsernameHash<'_>>,
+        e164: Option<Versioned<(E164, Vec<u8>)>>,
+        username_hash: Option<Versioned<UsernameHash<'_>>>,
         stored_account_data: Option<AccountData>,
         distinguished_tree_head: &LastTreeHead,
     ) -> Result<MaybePartial<AccountData>, RequestError<Error>> {
         let chat_search_response = self
             .chat
             .search(
-                aci,
+                aci.clone(),
                 aci_identity_key,
-                e164.as_ref(),
-                username_hash.as_ref(),
+                e164.as_ref().map(|x| x.as_ref()),
+                username_hash.as_ref().map(|x| x.as_ref()),
                 stored_account_data.as_ref(),
                 distinguished_tree_head,
             )
@@ -478,7 +625,7 @@ impl UnauthenticatedChatApi for KeyTransparencyClient<'_> {
         verify_chat_search_response(
             &self.inner,
             aci,
-            e164.map(|(e164, _)| e164),
+            e164.map(|x| x.map(|(e164, _)| e164)),
             username_hash,
             stored_account_data,
             chat_search_response,
@@ -683,7 +830,7 @@ impl UnauthenticatedChatApi for KeyTransparencyClient<'_> {
 
 fn verify_single_search_response(
     kt: &KeyTransparency,
-    search_key: Vec<u8>,
+    search_key: Versioned<Vec<u8>>,
     response: CondensedTreeSearchResponse,
     monitoring_data: Option<MonitoringData>,
     full_tree_head: &FullTreeHead,
@@ -692,7 +839,10 @@ fn verify_single_search_response(
     now: SystemTime,
 ) -> Result<VerifiedSearchResult, Error> {
     let result = kt.verify_search(
-        SlimSearchRequest::new(search_key),
+        SlimSearchRequest {
+            search_key: search_key.item,
+            version: search_key.version,
+        },
         FullSearchResponse::new(response, full_tree_head),
         SearchContext {
             last_tree_head,
@@ -707,9 +857,9 @@ fn verify_single_search_response(
 
 fn verify_chat_search_response(
     kt: &KeyTransparency,
-    aci: &Aci,
-    e164: Option<E164>,
-    username_hash: Option<UsernameHash>,
+    aci: Versioned<&Aci>,
+    e164: Option<Versioned<E164>>,
+    username_hash: Option<Versioned<UsernameHash>>,
     stored_account_data: Option<AccountData>,
     chat_search_response: TypedSearchResponse,
     last_distinguished_tree_head: Option<&LastTreeHead>,
@@ -742,7 +892,7 @@ fn verify_chat_search_response(
 
     let aci_result = verify_single_search_response(
         kt,
-        aci.as_search_key(),
+        aci.map(|x| x.as_search_key()),
         aci_search_response,
         aci_monitoring_data,
         &full_tree_head,
@@ -757,7 +907,7 @@ fn verify_chat_search_response(
                 .map(|(e164, e164_search_response)| {
                     verify_single_search_response(
                         kt,
-                        e164.as_search_key(),
+                        e164.map(|x| x.as_search_key()),
                         e164_search_response,
                         e164_monitoring_data,
                         &full_tree_head,
@@ -780,7 +930,7 @@ fn verify_chat_search_response(
             .map(|(username_hash, username_hash_response)| {
                 verify_single_search_response(
                     kt,
-                    username_hash.as_search_key(),
+                    username_hash.map(|x| x.as_search_key()),
                     username_hash_response,
                     username_hash_monitoring_data,
                     &full_tree_head,
@@ -897,6 +1047,11 @@ pub(crate) mod test_support {
         pub fn username_hash() -> UsernameHash<'static> {
             UsernameHash(Cow::Borrowed(USERNAME_HASH))
         }
+
+        // To be used for search function
+        pub fn e164_pair() -> (E164, Vec<u8>) {
+            (PHONE_NUMBER, UNIDENTIFIED_ACCESS_KEY.to_vec())
+        }
     }
 
     // Try connect/send operations to the real server this many times before failing
@@ -962,8 +1117,10 @@ pub(crate) mod test_support {
 #[cfg(test)]
 mod test {
     use std::cell::Cell;
+    use std::sync::{Arc, Mutex};
 
     use assert_matches::assert_matches;
+    use itertools::assert_equal;
     use prost::Message as _;
     use test_case::test_case;
 
@@ -1010,9 +1167,9 @@ mod test {
 
         let result = verify_chat_search_response(
             &kt,
-            &aci,
-            e164,
-            username_hash,
+            Versioned::from(&aci),
+            e164.map(Versioned::from),
+            username_hash.map(Versioned::from),
             Some(account_data),
             test_search_response(),
             Some(&test_distinguished_tree()),
@@ -1052,9 +1209,9 @@ mod test {
 
         let result = verify_chat_search_response(
             &kt,
-            &aci,
-            Some(e164),
-            Some(username_hash),
+            Versioned::from(&aci),
+            Some(Versioned::from(e164)),
+            Some(Versioned::from(username_hash)),
             Some(account_data),
             search_response,
             Some(&test_distinguished_tree()),
@@ -1066,16 +1223,31 @@ mod test {
         );
     }
 
+    #[derive(Default)]
+    struct SearchStub {
+        result: Option<Result<MaybePartial<AccountData>, RequestError<Error>>>,
+        invocations: Vec<SearchVersions>,
+    }
+
+    impl SearchStub {
+        fn new(res: Result<MaybePartial<AccountData>, RequestError<Error>>) -> Self {
+            Self {
+                result: Some(res),
+                invocations: vec![],
+            }
+        }
+    }
+
     struct TestKt {
         monitor: Cell<Option<Result<AccountData, RequestError<Error>>>>,
-        search: Cell<Option<Result<MaybePartial<AccountData>, RequestError<Error>>>>,
+        search: Arc<Mutex<SearchStub>>,
     }
 
     impl TestKt {
         fn for_monitor(monitor: Result<AccountData, RequestError<Error>>) -> Self {
             Self {
                 monitor: Cell::new(Some(monitor)),
-                search: Cell::new(None),
+                search: Arc::new(Mutex::new(Default::default())),
             }
         }
 
@@ -1085,7 +1257,7 @@ mod test {
         ) -> Self {
             Self {
                 monitor: Cell::new(Some(monitor)),
-                search: Cell::new(Some(search)),
+                search: Arc::new(Mutex::new(SearchStub::new(search))),
             }
         }
     }
@@ -1093,16 +1265,26 @@ mod test {
     impl UnauthenticatedChatApi for TestKt {
         fn search(
             &self,
-            _aci: &Aci,
+            aci: Versioned<&Aci>,
             _aci_identity_key: &PublicKey,
-            _e164: Option<(E164, Vec<u8>)>,
-            _username_hash: Option<UsernameHash<'_>>,
+            e164: Option<Versioned<(E164, Vec<u8>)>>,
+            username_hash: Option<Versioned<UsernameHash<'_>>>,
             _stored_account_data: Option<AccountData>,
             _distinguished_tree_head: &LastTreeHead,
         ) -> impl Future<Output = Result<MaybePartial<AccountData>, RequestError<Error>>> + Send
         {
-            let result = self.search.take().expect("unexpected call to search");
-            std::future::ready(result)
+            let mut guard = self.search.lock().expect("can lock");
+            guard.invocations.push(SearchVersions {
+                aci: aci.version,
+                e164: e164.and_then(|x| x.version),
+                username_hash: username_hash.and_then(|x| x.version),
+            });
+
+            if let Some(result) = guard.result.as_ref() {
+                std::future::ready(result.clone())
+            } else {
+                panic!("unexpected call to search")
+            }
         }
 
         fn distinguished(
@@ -1171,22 +1353,42 @@ mod test {
         assert_eq!(actual, monitor_result.into());
     }
 
+    #[derive(Clone, Copy)]
     enum BumpVersionFor {
         Aci,
         E164,
         UsernameHash,
     }
 
-    impl BumpVersionFor {
-        pub fn apply(&self, acc_data: &mut AccountData) {
-            let subject = match self {
-                BumpVersionFor::Aci => &mut acc_data.aci,
-                BumpVersionFor::E164 => acc_data.e164.as_mut().unwrap(),
-                BumpVersionFor::UsernameHash => acc_data.username_hash.as_mut().unwrap(),
+    trait Bumpable {
+        fn apply(&mut self, bump: BumpVersionFor);
+    }
+
+    impl Bumpable for AccountData {
+        fn apply(&mut self, bump: BumpVersionFor) {
+            let subject = match bump {
+                BumpVersionFor::Aci => &mut self.aci,
+                BumpVersionFor::E164 => self.e164.as_mut().unwrap(),
+                BumpVersionFor::UsernameHash => self.username_hash.as_mut().unwrap(),
             };
             // inserting a newer version of the subject
             let max_version = subject.greatest_version();
             subject.ptrs.insert(u64::MAX, max_version + 1);
+        }
+    }
+
+    impl Bumpable for SearchVersions {
+        fn apply(&mut self, bump: BumpVersionFor) {
+            let subject = match bump {
+                BumpVersionFor::Aci => &mut self.aci,
+                BumpVersionFor::E164 => &mut self.e164,
+                BumpVersionFor::UsernameHash => &mut self.username_hash,
+            };
+            if let Some(v) = subject.as_mut() {
+                *v += 1;
+            } else {
+                *subject = Some(1);
+            }
         }
     }
 
@@ -1196,7 +1398,7 @@ mod test {
     #[test_case(BumpVersionFor::UsernameHash; "newer username hash")]
     async fn monitor_and_search_e164_changed(bump: BumpVersionFor) {
         let mut monitor_result = test_account_data();
-        bump.apply(&mut monitor_result);
+        monitor_result.apply(bump);
 
         let kt = TestKt::new(
             Ok(monitor_result.clone()),
@@ -1226,17 +1428,86 @@ mod test {
     }
 
     #[tokio::test]
+    async fn monitor_and_search_self_monitor_fall_back_on_version_change() {
+        let mut monitor_result = test_account_data();
+        monitor_result.apply(BumpVersionFor::Aci);
+        monitor_result.apply(BumpVersionFor::E164);
+        monitor_result.apply(BumpVersionFor::E164);
+        monitor_result.apply(BumpVersionFor::UsernameHash);
+        monitor_result.apply(BumpVersionFor::UsernameHash);
+        monitor_result.apply(BumpVersionFor::UsernameHash);
+
+        let kt = TestKt::new(Ok(monitor_result.clone()), Ok(test_account_data().into()));
+
+        let result = monitor_and_search(
+            &kt,
+            &test_account::aci(),
+            &test_account::aci_identity_key(),
+            Some(test_account::e164_pair()),
+            Some(test_account::username_hash()),
+            test_account_data(),
+            &test_distinguished_tree(),
+            MonitorMode::MonitorSelf,
+        )
+        .await;
+
+        assert_matches!(result, Ok(actual) => {
+            // The final result after all the search-with-versions succeed
+            // should be what monitor originally returned.
+            assert_eq!(actual.inner, monitor_result);
+        });
+
+        assert_matches!(
+            kt.search.lock().expect("can lock").invocations[..],
+            [
+                SearchVersions {
+                    aci: Some(1),
+                    e164: None,
+                    username_hash: None,
+                },
+                SearchVersions {
+                    aci: Some(0),
+                    e164: Some(1),
+                    username_hash: None,
+                },
+                SearchVersions {
+                    aci: Some(0),
+                    e164: Some(2),
+                    username_hash: None,
+                },
+                SearchVersions {
+                    aci: Some(0),
+                    e164: None,
+                    username_hash: Some(1)
+                },
+                SearchVersions {
+                    aci: Some(0),
+                    e164: None,
+                    username_hash: Some(2)
+                },
+                SearchVersions {
+                    aci: Some(0),
+                    e164: None,
+                    username_hash: Some(3)
+                },
+            ]
+        );
+    }
+
+    #[tokio::test]
     #[test_case(BumpVersionFor::Aci; "newer Aci")]
     #[test_case(BumpVersionFor::E164; "newer E.164")]
     #[test_case(BumpVersionFor::UsernameHash; "newer username hash")]
-    async fn monitor_and_search_self_monitor_fail_on_version_change(bump: BumpVersionFor) {
+    async fn monitor_and_search_self_monitor_fall_back_search_with_version_failure(
+        bump: BumpVersionFor,
+    ) {
         let mut monitor_result = test_account_data();
-        bump.apply(&mut monitor_result);
+        monitor_result.apply(bump);
 
         let kt = TestKt::new(
             Ok(monitor_result.clone()),
             Err(RequestError::Unexpected {
-                log_safe: "pass through unexpected error".to_owned(),
+                log_safe: "search with version failed".to_string(),
             }),
         );
 
@@ -1244,19 +1515,27 @@ mod test {
             &kt,
             &test_account::aci(),
             &test_account::aci_identity_key(),
-            None,
-            None,
+            Some(test_account::e164_pair()),
+            Some(test_account::username_hash()),
             test_account_data(),
             &test_distinguished_tree(),
             MonitorMode::MonitorSelf,
         )
         .await;
 
-        // monitor invocation should have succeeded, and search
-        // should not have been invoked
-        assert_matches!(
-            result,
-            Err(RequestError::Other(Error::VerificationFailed(_)))
+        assert_matches!(result, Err(RequestError::Unexpected { .. }));
+
+        let mut expected_versions = SearchVersions {
+            // ACI is required, so we will have to search for it always.
+            aci: Some(0),
+            e164: None,
+            username_hash: None,
+        };
+        expected_versions.apply(bump);
+
+        assert_equal(
+            kt.search.lock().expect("can lock").invocations.clone(),
+            [expected_versions],
         );
     }
 
@@ -1302,6 +1581,36 @@ mod test {
         assert_eq!(
             search_result_account_data,
             updated_account_data.into_inner()
+        );
+    }
+
+    #[tokio::test]
+    #[test_case(BumpVersionFor::Aci; "newer Aci")]
+    #[test_case(BumpVersionFor::E164; "newer E.164")]
+    #[test_case(BumpVersionFor::UsernameHash; "newer username hash")]
+    async fn monitor_and_search_self_monitor_version_jump_too_big(bump: BumpVersionFor) {
+        let mut monitor_result = test_account_data();
+        for _ in 0..MAXIMUM_ALLOWED_VERSION_DELTA + 1 {
+            monitor_result.apply(bump);
+        }
+
+        let kt = TestKt::for_monitor(Ok(monitor_result));
+
+        let result = monitor_and_search(
+            &kt,
+            &test_account::aci(),
+            &test_account::aci_identity_key(),
+            Some(test_account::e164_pair()),
+            Some(test_account::username_hash()),
+            test_account_data(),
+            &test_distinguished_tree(),
+            MonitorMode::MonitorSelf,
+        )
+        .await;
+
+        assert_matches!(
+            result,
+            Err(RequestError::Other(Error::VerificationFailed(_)))
         );
     }
 }